chore: reorganize tests to match the modules
